Adjust SelectedTapMobile and SVG imports. Minor fixes
- Add PHLASK statement props in SelectedTapMobile instead of hardcoded
  "N/A"
- Remove SelectedTapMobile and rename SelectedTapMobileInfo to
  SelectedTapMobile
- Add SwipeableDrawer to SelectedTap.js
- Instead of using hard-coded <svg> elements in a separate file, it now uses
  {ReactComponent as [filename]}  to import svg files
- Remove IconBtn and Icon components
